[main-info] -Changed some lines of code from main.js file and deleted not relating pictures
