+++ src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Main from "./Main/Main";
+import Main from "./components/Main/Main";
 
 function App() {
   return (
